#602 | chore(masonry): rename path utility to brick-shape
